+++ gumption/src/app/app.component.html
@@ -7,10 +7,6 @@
       <div class="col-md-10">
         <div class="row controls-container">
           <div>
-            <app-count></app-count>
-          </div>
-          <div>
-            <app-pagination></app-pagination>
           </div>
           <div>
               <app-sort></app-sort>
