+++ gumption/src/app/results/count/count.component.html
