+++ gumption/src/app/results/results.component.html
@@ -1,3 +1,4 @@
-<div *ngFor="let result of results">
-    <app-result [result]="result"></app-result>
-</div>
+<span style="display:inline-block">
+  <app-count></app-count>
+  <app-pagination></app-pagination>
+</span>
