+++ gumption/src/app/shared/services/url-resolver.service.ts
@@ -2,22 +2,25 @@ import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/take';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { Router, Resolve, RouterStateSnapshot,
-         ActivatedRouteSnapshot } from '@angular/router';
+import {
+  Router, Resolve, RouterStateSnapshot,
+  ActivatedRouteSnapshot
+} from '@angular/router';
 import { Query } from '../interfaces/query';
 import { QueryService } from '../services/query.service';
 
 @Injectable()
-export class UrlResolver implements Resolve<Object[]> {
-  constructor(private router: Router,
-              private queryService: QueryService) {}
+export class UrlResolver implements Resolve<object[]> {
+  constructor(
+    private router: Router,
+    private queryService: QueryService) { }
 
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<Object[]> {
+  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<object[]> {
     const query: Query = {
-      queryString : route.queryParams['q'] || '',
-      page : route.queryParams['page'] || null,
-      sortBy : route.queryParams['sortBy'] || null,
-      sortOrder : route.queryParams['order'] || null
+      queryString: route.queryParams['q'] || '',
+      page: route.queryParams['page'] || null,
+      sortBy: route.queryParams['sortBy'] || null,
+      sortOrder: route.queryParams['order'] || null
     };
     return this.queryService.fetchResults(query);
   }
